Add wsgi compatible app class
This adds a basic app class, request processor and request class.
It uses the router to find the appropriate function to call then yields/returns the appropriate things at each stage according to the wsgi spec.

The RequestProcessor encapsulates handling the actual request (i.e creating a request object and delegating to the router)

The RequestProcessor encapsulates handling the actual request (i.e creating a request object and delegating to the router).

The Request class wraps the input stream and will expose methods to read data from it.